board: mediatek: Add CMF Phone 1 (tetris) support
The CMF Phone 1 is a midrange Android phone released by Nothing
Technology in 2024. Add support for it, when booting u-boot as lk.

NOTE: CONFIG_OF_EMBED is enabled because of the specific boot process to
u-boot that this device uses. We cannot pass a dtb to u-boot due to
limitations of how this device loads and launches u-boot (simply jumps
to the address from TEE, which cannot be modified)

Signed-off-by: Igor Belwon <igor.belwon@mentallysanemainliners.org>